renamed routes.js to requestHandler.js
